Adjust tests to get around obscure Rails bug
There is an obscure bug in Rails that causes DB constraint errors to be raised
when assigning directly to a has_many :through association on an STI model that
has accepts_nested_attributes_for set on the join model and DB constraints on
the foreign keys of the join model[1]. This kind of assignment only ever happens
in the test suite (in the running app, topics are assigned to organisations in
the admin interface by creating join instances (organisation_classifications)
rather than assigning to the has_many :through).
Rather than add a temporary hack to the assignment method until the bug is fixed
(there is an open pull request), it makes more sense to adjust the tests so they
don't use the broken association assignment method.

[1] https://github.com/rails/rails/pull/18198